+++ coursework/REITSim/bin/Debug/net7.0/GUI/Mediator.py
@@ -1,5 +1,5 @@
 from tkinter import *
-from GUIComponents import *
+from GUI.GUIComponents import *
 
 
 class Mediator:
