+++ coursework/REITSim/bin/Debug/net7.0/GUI/PageManager.py
@@ -1,5 +1,5 @@
 from tkinter import *
-from Mediator import *
+from GUI.Mediator import *
 
 
 class PageManager:
